Add a system tray icon.
Make it so that when user clicks the X in the MainWindow, that the
program will simply 'minimize' to tray.  The alarm should keep the state
(running or not) when user closes MainWindow.

Make it so that user can stop the alarm by right clicking the system
tray.

Restore MainWindow on single click of tray icon.

Hide MainWindow on double click of tray icon.