Remove and replace the statusbar with a label so we can have the text centered in the UI.
Add a message box section to the UI.  Now user can type some text that
will remind them of what this timer is for.  This will also allow the
user to make sense of multiple instances of this applicatin running.

Add code so that once the user started the timer, the minutes and
message section cannot be changed until the timer is stopped.

Add the message to the applications system tray icon, again to aid in
multiple running instance identification.

Add the message to the pop up notification dialog.

Since QTextEdit does not have a 'on return key pressed' event, implement
event filter to grab this event and start the timer.

Pressing enter on the minutes input field now advanced user to the
message box entry field.

Rearrange the UI in designer.

Minor code cleanup.