Remove the stop button.
Make the start button a toggle to start/stop

Refactor the start buttons method to be a toggle.

Remove CSS for hovers in designer and put them in local method that will
change the hover color of the button using an enum of green or red.

Rearrange the GUI to look nice with just one button.

Remove old stale comments.