Move items for the ui out of the ctor and put into a new tweakUi function.
Impl css theme for the system tray.  The tray and line edit and text
edit have inconsistent behavior between 5.12 and 5.14 Qt, and this is
lesser of two evils to make it behave on both versions.

Make app exit on window close for now until we release, add TODO for
same.