Add sound file picker wizard and a config file in user's Roaming dir.
config file should store the users sound file preference.

main.cpp: add QSettings and pass the vars to mainwindow.cpp

mainwindow.cpp:
- add QSettings support.  Modify initialization obj creation so we get
passed the vars from main.cpp.
- override contextMenuEvent so we can have a custom context menu when
user clicks on app window background (outside of the input fields) so as
to run the runSoundFilePickerWizard()
- add runSoundFilePickerWizard - bring up the wizard in new window,
leave timer's main window running.
- add palette WindowText so QDialog of the wizard will inherit.

Add new wizard class to allow user to pick new sound file.  Reset it
back to "FACTORY" if the reject or cancel. Close and clean up the wizard
when they're done.

This is not feature complete and ready for production.